#166 refactor: userAssembler 함수 -> DTO, Entity 이전
