+++ src/main/kotlin/com/psr/psr/user/dto/phoneReq/SMSReq.kt
@@ -1,5 +1,7 @@
 package com.psr.psr.user.dto.phoneReq
 
+import com.psr.psr.user.dto.request.ValidPhoneReq
+
 data class SMSReq (
     val type: String?= "SMS",
     val contentType: String?="COMM",
@@ -7,4 +9,15 @@ data class SMSReq (
     val from: String,
     val content: String,
     val messages: List<MessageReq>
-)
+){
+    companion object{
+        fun toSMSReqDto(validPhoneReq: ValidPhoneReq, key: String, sendPhone: String) : SMSReq{
+            val message = MessageReq(to = validPhoneReq.phone)
+            return SMSReq(
+                from = sendPhone,
+                content = "[PSR] 인증번호는 [ $key ] 을 입력해주세요",
+                messages = listOf(message) // 싱글톤 list = Collections.singletonList()
+            )
+        }
+    }
+}
