+++ src/main/kotlin/com/psr/psr/user/dto/response/MyPageInfoRes.kt
@@ -1,5 +1,7 @@
 package com.psr.psr.user.dto.response
 
+import com.psr.psr.user.entity.User
+
 
 data class MyPageInfoRes(
     val email: String,
@@ -7,4 +9,10 @@ data class MyPageInfoRes(
     val type: String,
     val phone: String,
     val nickname: String,
-)
+){
+    companion object{
+        fun toMyPageInfoRes(user: User) : MyPageInfoRes {
+            return MyPageInfoRes(user.email, user.imgUrl, user.type.value, user.phone, user.nickname)
+        }
+    }
+}
